--- src/air/__main__.py
@@ -1,3 +1,0 @@
-import air
-
-air.main()
