+++ src/air/plugins/markdown_plugin.py
@@ -1,5 +1,4 @@
 import markdown
-from pathlib import Path
 from air.generator import Plugin
