+++ tests/test_main.py
@@ -1,3 +1,5 @@
+from air.cli import main
+
+
 def test_main():
-    from air import main
     assert main() == 0
